:+1:  [FIX] --> Classes and code refactor
Classes and code refactor
