+++ app/src/main/java/com/globant/samples/volley/data/remote/UserApiService.java
@@ -27,13 +27,13 @@ public interface UserApiService {
     @GET(ApiConstants.BASE_URL_REPOS + "{github_user_name}/repos")
     Observable<List<GithubUserRepo>> doGetUserRepos(@Path("github_user_name") String userName);
 
-    class Factory {
-        public static UserApiService create() {
-            Gson gson = new GsonBuilder().setDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'").create();
-            Retrofit retrofit = new Retrofit.Builder().baseUrl(ApiConstants.BASE_URL)
-                    .addConverterFactory(GsonConverterFactory.create(gson))
-                    .addCallAdapterFactory(RxJavaCallAdapterFactory.create()).build();
-            return retrofit.create(UserApiService.class);
-        }
-    }
+//    class Factory {
+//        public static UserApiService create() {
+//            Gson gson = new GsonBuilder().setDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'").create();
+//            Retrofit retrofit = new Retrofit.Builder().baseUrl(ApiConstants.BASE_URL)
+//                    .addConverterFactory(GsonConverterFactory.create(gson))
+//                    .addCallAdapterFactory(RxJavaCallAdapterFactory.create()).build();
+//            return retrofit.create(UserApiService.class);
+//        }
+//    }
 }
