+++ app/src/main/java/com/globant/samples/volley/data/remote/factory/Factory.java
@@ -1,0 +1,25 @@
+package com.globant.samples.volley.data.remote.factory;
+
+import com.globant.samples.volley.data.remote.ApiConstants;
+import com.globant.samples.volley.data.remote.UserApiService;
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+
+import retrofit2.Retrofit;
+import retrofit2.adapter.rxjava.RxJavaCallAdapterFactory;
+import retrofit2.converter.gson.GsonConverterFactory;
+
+/**
+ * Created by miller.barrera on 23/06/2017.
+ */
+
+public class Factory {
+
+    public static UserApiService create() {
+        Gson gson = new GsonBuilder().setDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'").create();
+        Retrofit retrofit = new Retrofit.Builder().baseUrl(ApiConstants.BASE_URL)
+                .addConverterFactory(GsonConverterFactory.create(gson))
+                .addCallAdapterFactory(RxJavaCallAdapterFactory.create()).build();
+        return retrofit.create(UserApiService.class);
+    }
+}
