+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Volley Sample</string>
+    <string name="app_name">RxJavaSample</string>
 </resources>
